Updated CmakeLists to build for VS Code
Removed usage of $<CONFIG> because AFAIK only the Visual Sudio IDE
generates Debug/Release build directories by default.
I must investigate more on how this can be solved together with CMake.

Also added usage of WIN32 to executable. This will ensure that we
bild an executable with a WinMain entry point on windows.